server/requirement.txt
alembic==1.13.1
aniso8601==9.0.1
attrs==23.2.0
bidict==0.23.1
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
flask-restx==1.3.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
flask-swagger-ui==4.11.1
greenlet==3.0.3
h11==0.14.0
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
Mako==1.3.5
MarkupSafe==2.1.5
PyJWT==2.8.0
python-dotenv==1.0.1
python-engineio==4.9.0
python-socketio==5.11.2
pytz==2024.1
referencing==0.35.1
rpds-py==0.18.1
simple-websocket==1.0.0
SQLAlchemy==2.0.30
typing_extensions==4.11.0
Werkzeug==3.0.3
wsproto==1.2.0
